Added dashboard caching, redesigned login to be a global modal, updated the way authentication works

+++ backend/app/models/pricing_model.py
@@ -2,5 +2,5 @@ A = 1.35
 B = 0.11282
 
 
-def price_model(lp: int) -> float:
+def price_model(lp: float) -> float:
     return (lp ** A) * B
